requirements.txt
qrcode
python-barcode

